+++ Changelog.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add feature at compmgr engine which provide static server route to support honojs. Completed 2025-10-22
 - Add module loading method which support both frontend and backend. Utils selected modules will apply it. Completed 2025-10-25
 - Implement `honoassist` module to webnodehonojs engine instead curent render method. COmpleted on 2025-10-26
+- The repetitive modules from both webbunjs and webnodehonojs engine move to compmgr/hono to standardise the web render procedure. Completed on 2025-10-26
 
 ### Changed
 
@@ -34,7 +35,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Remove duplicate code from onrequest function as `reaction.js` in webbunjs and webnodehonojs engine. Completed on 2025-10-22
 - Remove `@sagold/json-pointer` dependencies from main package.json. Completed on 2025-10-22
 - Remove `jptr` the module from global sysmodule at `app.js`. Completed on 2025-10-22
-- Renove unused function and code from webbunjs engine. Completed on 2025-10-26
+- Remove unused function and code from webbunjs engine. Completed on 2025-10-26
 
 ### Fixed